Move bot code -> bot.js
